Use savannah.gnu.org address to download GPG keys
It appears that there is a certificate issue with https://sv.gnu.org which prevents this action from installing Guix successfully as it fails at the "Fetch key" step.  This change merely uses the full savannah.gnu.org address as it doesn't seem to be affected by the certificate issue.